Revert "Merge pull request #22 from kirill-dorkin/add-geolocation-based-language-localization"
This reverts commit afcf7ba9e6abe61efbf68d04e25c271b3432079f, reversing
changes made to 6d19e2b06647809b825921653ca89b38b9fd5f30.
