--- i18n.ts
@@ -1,5 +1,0 @@
-export const locales = ['ru', 'ky', 'en'] as const;
-export const defaultLocale = 'ru';
-// Always require a locale prefix in the URL to avoid redirect loops
-export const localePrefix = 'always';
-
