+++ package.json
@@ -35,7 +35,6 @@
     "mongoose": "^8.9.2",
     "next": "15.0.2",
     "next-auth": "^4.24.11",
-    "next-intl": "4.1.0",
     "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "18.3.1",
