--- src/app/[locale]/page.tsx
@@ -1,9 +1,0 @@
-import Hero from "@/components/home/Hero";
-
-export default function Home() {
-  return (
-    <>
-      <Hero />
-    </>
-  );
-}
