+++ src/app/admin/devices/page.tsx
