+++ src/components/home/Hero.tsx
@@ -5,13 +5,11 @@ import { Button } from "@/components/ui/button";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useSession } from "next-auth/react";
-import { useTranslations } from "next-intl";
 
 const Hero: React.FC = () => {
   const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
   const { data: session } = useSession();
   const router = useRouter();
-  const t = useTranslations("Hero");
 
   useEffect(() => {
     if (session?.user) return setIsLoggedIn(true);
@@ -38,21 +36,22 @@ const Hero: React.FC = () => {
     >
       <div className="relative flex flex-col items-center text-center px-6 z-10">
         <h1 className="text-4xl sm:text-5xl md:text-6xl font-bold">
-          {t("title")}
+          Best Electronics Service System
         </h1>
         <p className="mb-8 max-w-xl">
-          {t("description")}
+          We provide professional electronics repairs and upgrades. Use the request
+          button below to submit a service form.
         </p>
         <div className="flex flex-col sm:flex-row justify-center gap-4">
           {!isLoggedIn && (
             <Button size="lg" asChild className="w-fit">
-              <Link href="/request">{t("request")}</Link>
+              <Link href="/request">Request Repair</Link>
             </Button>
           )}
           {isLoggedIn && (
             <>
               <Button size="lg" className="w-fit" onClick={handleGoToDashboard}>
-                {t("dashboard")}
+                Go to Dashboard
               </Button>
               <Button
                 size="lg"
@@ -60,7 +59,7 @@ const Hero: React.FC = () => {
                 className="w-fit"
                 onClick={handleLogout}
               >
-                {t("logout")}
+                Logout
               </Button>
             </>
           )}
