--- src/messages/ky.json
@@ -1,15 +1,0 @@
-{
-  "Hero": {
-    "title": "Best Electronics тейлөө системасы",
-    "description": "Биз электрониканы кесипкөй оңдоо жана жаңыртууларды сунуштайбыз. Өтүнмө калтыруу үчүн төмөнкү баскычты басыңыз.",
-    "request": "Өтүнмө калтыруу",
-    "dashboard": "Дашбордго өтүү",
-    "logout": "Чыгуу"
-  },
-  "Login": {
-    "email": "Email",
-    "password": "Сырсөз",
-    "button": "Кирүү",
-    "error": "Кирүү маалыматтары туура эмес"
-  }
-}
