--- src/messages/ru.json
@@ -1,15 +1,0 @@
-{
-  "Hero": {
-    "title": "Система сервиса Best Electronics",
-    "description": "Мы предлагаем профессиональный ремонт и модернизацию электроники. Нажмите кнопку заявки ниже, чтобы оставить заявку.",
-    "request": "Оставить заявку",
-    "dashboard": "Перейти в панель",
-    "logout": "Выйти"
-  },
-  "Login": {
-    "email": "Email",
-    "password": "Пароль",
-    "button": "Войти",
-    "error": "Неверные учетные данные"
-  }
-}
